Segmented code and handle config better Configuration for all services is now contained within ScoreboardState, where all the services themselves are also held.
The majority of code has also been segmented out into individual files
to make management easier.

Service checking is now implemented and service checking is very stable.
Services that are slow, or take a long time to respond such as SMTP
don't score as well as quick services like HTTP but this is due to the
nature of those particular services, not this program. The solution for
this issue is for the user to set the service timeout config option as
needed until a service they are sure is operational is scored properly.