Actually working host-command. Reworked startup Changed the service startup from waiting before querying to immediately querying on startup. This prevents us from having to worry about default state of services before they can be tested.
Corrected host-command procedures to actually work correctly. The
problem was that variables generated from range that return something
other than the arrays index get locked when passed directly to an
asyncronous function. This is by design (I think). The way around this
is to directly access the array element by using an index.

Edited the buildConfigFunction to generate an actually working config
service wise.
accessing