Scoreboard web interface. Analytics. Cleaned up. Added a web interface with a default. TODO: Add support to specify a directory containing a user-customizable html template for the scoreboard.
Added service uptime and downtime tracking. Data is not displayed in a
final form, but is being recorded. Not sure how I wan't to display it
yet, as that data isn't meant for contestants.

Cleaned up the scoreboardStateUpdater so that it only looks through
services and hosts until it finds what it needs, not after.