Cleaned up web scoreboard. Better timing, new docs Implemented a web based scoreboard that by default shows nice information about services including their overall up state, and duration of uptime and duration of downtime. The webpage automatically refreshes every 5 seconds to get new content from the webserver. In the future, javascript should be written to check for updates behind the scenes and keep the timers running, so that the backend of the webserver can be re-written to only update the web content when a service state has been changed.
Service uptime and downtime tracking has been re-done to be better
suited and show better information, as well as extended debug messages
in the console which include this information.

Re-wrote a lot of docs, changed access modifiers and wrote docs for new
functions and methods.

Updated the help page to be more helpful :D

re-segmented code appropriately

re-wrote func main() to be shorter and make more sense. The entire
scoreboard is started via the Start() method

Implemented new config options to bind the scoreboard to a specific
address, set a finite duration to the testing of services and hosts -
and as a result all the supporting threads for those functions such as
the state updater and webcontent updater. Also added a way for users to
specify their own scoreboard file in the config file. Documentation will
need to be written about the template and data users can access in the
templates so users can create their own scoreboard.

Threads that run to update portions of the scoreboard, or test services
and hosts now have a means by which they can shut down after the newly
configured time.