Updated websheet updating through templates The WebContentUpdater now only locks the scoreboard content when an update has occured. This is triggered by a golang channel.
A future modification might be to create sub templates for timers?
By doing this we might save some compute power on regenerating
the entire web content. We might not though, and this would just
be a feel good change. If timers are segmented to a subtemplate,
then the correct place to execute those subtemplates would be in
ServeHTTP. This is probably the correct method to generate up to date
timers without having to majorly control the data from State.

Currently, timers are updated by regenerating the entire webSheet with a
saved template and saved data. We don't save on generating the template
or on data, but we do save significantly on read locking the scoreboard
State.